+++ bin/library_manager.js
@@ -1,6 +1,3 @@
-/**
- * Created by mkrooted on 3/9/2017.
- */
 const db = require("./database");
 const fs = require("fs");
 const path = require("path");
@@ -30,14 +27,16 @@ function loadLibraries(callback) {
             });
         },
         (res, cb) => {
-            const path_base = path.join(__dirname, "../../hub_data/libraries");
+            const path_base = path.join(__dirname, "../hub_data/libraries");
             const libs = getDirectories(path_base);
             for (let lib in libs) {
                 const str_path = path.join(path_base, libs[lib]);
                 const library_data = JSON.parse(fs.readFileSync(str_path + "/library.json"));
                 db.models.Library.add({
                     "lib_name": library_data.name,
+                    "lib_interface": library_data.interface,
                     "lib_uids": JSON.stringify(library_data.supported_devices),
+                    "lib_title": library_data.title
                 }, function (err, data) {
                     if (err) {
                         logger.error(TAG, err);
@@ -48,7 +47,8 @@ function loadLibraries(callback) {
                     for (let func in functions_names) {
                         db.models.Function.add({
                             "function_library": lib_id,
-                            "function_name": func
+                            "function_name": func,
+                            "function_type": library_data.functions[func].type
                         }, function (err, data) {
                             if (err) {
                                 logger.error(TAG, err);