+++ bin/main_loop.js
@@ -1,4 +1,0 @@
-/**
- * Created by mkrooted on 3/11/2017.
- */
-