+++ configs/lan.js
@@ -1,6 +1,6 @@
 module.exports = {
-    network_address: "192.168.0.0",
+    network_address: "192.168.1.28",
     mask_bits: 24,
-    default_gateway: "192.168.0.1",
+    default_gateway: "192.168.1.1",
     scan_query: "192.168.1.0/24"
 };