+++ hub_data/interfaces/wifi/functions.js
@@ -1,16 +1,13 @@
-/**
- * Created by mkrooted on 2/21/2017.
- */
-
 var http = require("http");
 /*  Events:
  *   - GET_json_response (requestId, body)
  * */
 
-function GET_json(device_ip, path, callback) {
+function GET_json(device_ip, port, path, callback) {
     http.request({
         "hostname": device_ip,
-        "path": path
+        "path": path,
+        "port": port
     }, function (res) {
         var str = '';
 
@@ -25,13 +22,12 @@ function GET_json(device_ip, path, callback) {
         });
     });
 }
-function GET_string(device_ip, path, callback) {
-    console.log("I'm here");
-
+function GET_string(device_ip, port, path, callback) {
     var req = http.request({
         "hostname": device_ip,
         "path": path,
         "method": "GET",
+        "port": port
     }, function (res) {
         console.log("statuscode: "+res);
         var str = '';
@@ -56,5 +52,34 @@ function GET_string(device_ip, path, callback) {
     req.end();
 }
 
+function POST_json(device_ip, port, path, data, callback) {
+    var req = http.request({
+        "hostname": device_ip,
+        "path": path,
+        "method": "POST",
+        "port": port
+    }, function (res) {
+        var str = '';
+
+        //another chunk of data has been recieved, so append it to `str`
+        res.on('data', function (chunk) {
+            str += chunk;
+        });
+
+        //the whole response has been recieved, so we just print it out here
+        res.on('end', function () {
+            callback(str);
+        });
+    });
+
+    req.on('error', (e) => {
+        console.log('problem with request:', e);
+    });
+
+    req.write(data);
+    req.end();
+}
+
 module.exports.GET_json = GET_json;
 module.exports.GET_string = GET_string;
+module.exports.POST_json = POST_json;