+++ hub_data/libraries/led_library/functions.js
@@ -1,22 +1,17 @@
-/**
- * Created by mkrooted on 2/21/2017.
- */
-
-
 module.exports.hello = function (hard_interface, device, callback) {
-    hard_interface.GET_json(device.dev_address, "/hello", callback);
+    hard_interface.GET_json(device.dev_address, 80, "/hello", callback);
 };
 
 module.exports.get_status = function (hard_interface, device, callback) {
-    hard_interface.GET_string(device.dev_address, "/get_status", function (body) {
+    hard_interface.GET_string(device.dev_address, 80, "/get_status", function (body) {
         callback((body == "1"));
     });
 };
 
 module.exports.light_on = function (hard_interface, device, callback) {
-    hard_interface.GET_string(device.dev_address, "/light_on", callback);
+    hard_interface.GET_string(device.dev_address, 80, "/light_on", callback);
 };
 
 module.exports.light_off = function (hard_interface, device, callback) {
-    hard_interface.GET_string(device.dev_address, "/light_off", callback);
+    hard_interface.GET_string(device.dev_address, 80, "/light_off", callback);
 };